Added in life attributes and multiple types
I've added a decent chunk of complexity to the code by giving all of the pixels attributes.  Basically I gave them some attack, defense, and strength characteristics along with the propensity to do each.  The idea was to have a neutral, aggressor, defender, and grass type life forms.  It's getting a bit crowded in the code so I will have to refactor soon.  It's starting to look a bit better, but I need to have the other life forms treat grass more similarly to dead spaces for it to blossom a bit more.